Basic Server structure set up
